Refactor terrain module
