+++ emergence_lib/src/tiles/terrain.rs
@@ -1,6 +1,6 @@
 //! The [`TerrainTilemap`] manages visualization of terrain.
 
-use crate::terrain::TerrainType;
+use crate::terrain::terrain_types::TerrainType;
 use bevy::prelude::Component;
 use bevy_ecs_tilemap::map::TilemapTileSize;
 use indexmap::{indexmap, IndexMap};
